RAG_QA/STransformers RAG/requirements.txt
sentence-transformers 
faiss-cpu 
pypdf

llama-cpp-python
